fix(ios): centralize filename checks and harden path resolution
